requirements.txt
onnxruntime==1.16.3
numpy<2.0
pillow
requests
